+++ gdm/config.py
@@ -6,117 +6,10 @@ import logging
 import yorm
 
 from . import common
-from .shell import ShellMixin, GitMixin
+from . import shell
+from .source import Source
 
 log = logging.getLogger(__name__)
-
-
-@yorm.attr(repo=yorm.converters.String)
-@yorm.attr(dir=yorm.converters.String)
-@yorm.attr(rev=yorm.converters.String)
-@yorm.attr(link=yorm.converters.String)
-class Source(yorm.converters.AttributeDictionary, ShellMixin, GitMixin):
-    """A dictionary of `git` and `ln` arguments."""
-
-    def __init__(self, repo, dir, rev='master', link=None):  # pylint: disable=W0622
-        super().__init__()
-        self.repo = repo
-        self.dir = dir
-        self.rev = rev
-        self.link = link
-        if not self.repo:
-            raise ValueError("'repo' missing on {}".format(repr(self)))
-        if not self.dir:
-            raise ValueError("'dir' missing on {}".format(repr(self)))
-
-    def __repr__(self):
-        return "<source {}>".format(self)
-
-    def __str__(self):
-        fmt = "'{r}' @ '{v}' in '{d}'"
-        if self.link:
-            fmt += " <- '{s}'"
-        return fmt.format(r=self.repo, v=self.rev, d=self.dir, s=self.link)
-
-    def __eq__(self, other):
-        return self.dir == other.dir
-
-    def __ne__(self, other):
-        return self.dir != other.dir
-
-    def __lt__(self, other):
-        return self.dir < other.dir
-
-    def update_files(self, force=False, clean=True):
-        """Ensure the source matches the specified revision."""
-        log.info("Updating source files...")
-
-        # Enter the working tree
-        if not os.path.exists(self.dir):
-            log.debug("Creating a new repository...")
-            self.git_clone(self.repo, self.dir)
-        self.cd(self.dir)
-
-        # Check for uncommitted changes
-        if not force:
-            log.debug("Confirming there are no uncommitted changes...")
-            if self.git_changes():
-                common.show()
-                msg = "Uncommitted changes: {}".format(os.getcwd())
-                raise RuntimeError(msg)
-
-        # Fetch the desired revision
-        self.git_fetch(self.repo, self.rev)
-
-        # Update the working tree to the desired revision
-        self.git_update(self.rev, clean=clean)
-
-    def create_link(self, root, force=False):
-        """Create a link from the target name to the current directory."""
-        if self.link:
-            log.info("Creating a symbolic link...")
-            target = os.path.join(root, self.link)
-            source = os.path.relpath(os.getcwd(), os.path.dirname(target))
-            if os.path.islink(target):
-                os.remove(target)
-            elif os.path.exists(target):
-                if force:
-                    self.rm(target)
-                else:
-                    common.show()
-                    msg = "Preexisting link location: {}".format(target)
-                    raise RuntimeError(msg)
-            self.ln(source, target)
-
-    def identify(self, allow_dirty=True):
-        """Get the path and current repository URL and hash."""
-        if os.path.isdir(self.dir):
-
-            self.cd(self.dir)
-
-            path = os.getcwd()
-            url = self.git_get_url()
-            if self.git_changes(visible=True):
-                revision = '<dirty>'
-                if not allow_dirty:
-                    common.show()
-                    msg = "Uncommitted changes: {}".format(os.getcwd())
-                    raise RuntimeError(msg)
-            else:
-                revision = self.git_get_sha()
-            common.show(revision, log=False)
-
-            return path, url, revision
-
-        else:
-
-            return os.getcwd(), '<missing>', '<unknown>'
-
-    def lock(self):
-        """Return a locked version of the current source."""
-        _, _, sha = self.identify()
-        source = self.__class__(self.repo, self.dir, sha, self.link)
-        return source
 
 
 @yorm.attr(all=Source)
@@ -128,7 +21,7 @@ class Sources(yorm.converters.SortedList):
 @yorm.attr(sources=Sources)
 @yorm.attr(sources_locked=Sources)
 @yorm.sync("{self.root}/{self.filename}")
-class Config(ShellMixin):
+class Config:
     """A dictionary of dependency configuration options."""
 
     FILENAMES = ('gdm.yml', 'gdm.yaml', '.gdm.yml', '.gdm.yaml')
@@ -152,15 +45,16 @@ class Config(ShellMixin):
         return os.path.join(self.root, self.location)
 
     def install_deps(self, *names, depth=None,
-                     update=True, recurse=False, force=False, clean=True):
+                     update=True, recurse=False,
+                     force=False, fetch=False, clean=True):
         """Get all sources."""
         if depth == 0:
             log.info("Skipped directory: %s", self.location_path)
             return 0
 
         if not os.path.isdir(self.location_path):
-            self.mkdir(self.location_path)
-        self.cd(self.location_path)
+            shell.mkdir(self.location_path)
+        shell.cd(self.location_path)
 
         sources = self._get_sources(use_locked=False if update else None)
         dirs = list(names) if names else [source.dir for source in sources]
@@ -175,7 +69,7 @@ class Config(ShellMixin):
                 log.info("Skipped dependency: %s", source.dir)
                 continue
 
-            source.update_files(force=force, clean=clean)
+            source.update_files(force=force, fetch=fetch, clean=clean)
             source.create_link(self.root, force=force)
             count += 1
 
@@ -189,11 +83,12 @@ class Config(ShellMixin):
                     update=update and recurse,
                     recurse=recurse,
                     force=force,
+                    fetch=fetch,
                     clean=clean,
                 )
                 common.dedent()
 
-            self.cd(self.location_path, visible=False)
+            shell.cd(self.location_path, _show=False)
 
         common.dedent()
         if dirs:
@@ -204,7 +99,7 @@ class Config(ShellMixin):
 
     def lock_deps(self, *names, obey_existing=True):
         """Lock down the immediate dependency versions."""
-        self.cd(self.location_path)
+        shell.cd(self.location_path)
         common.show()
         common.indent()
 
@@ -227,7 +122,7 @@ class Config(ShellMixin):
 
             common.show()
 
-            self.cd(self.location_path, visible=False)
+            shell.cd(self.location_path, _show=False)
 
         if count:
             yorm.update_file(self)
@@ -235,13 +130,13 @@ class Config(ShellMixin):
 
     def uninstall_deps(self):
         """Remove the sources location."""
-        self.rm(self.location_path)
+        shell.rm(self.location_path)
         common.show()
 
     def get_deps(self, depth=None, allow_dirty=True):
         """Yield the path, repository URL, and hash of each dependency."""
         if os.path.exists(self.location_path):
-            self.cd(self.location_path)
+            shell.cd(self.location_path)
             common.show()
             common.indent()
         else:
@@ -265,7 +160,7 @@ class Config(ShellMixin):
                 )
                 common.dedent()
 
-            self.cd(self.location_path, visible=False)
+            shell.cd(self.location_path, _show=False)
 
         common.dedent()