+++ gdm/plugin.py
@@ -5,12 +5,12 @@
 import argparse
 import logging
 
-from . import PLUGIN, __version__
+from . import PLUGIN, NAME, __version__
 from . import common
 from .cli import _get_command, _run_command
 
 PROG = 'git ' + PLUGIN
-DESCRIPTION = "Use GDM (v{}) to manage source dependencies.".format(__version__)
+DESCRIPTION = "Use {} (v{}) to install repostories.".format(NAME, __version__)
 
 log = logging.getLogger(__name__)
 
@@ -66,6 +66,7 @@ def main(args=None):
     namespace.root = None
     namespace.depth = None
     namespace.allow_dirty = True
+    namespace.fetch = True
 
     # Configure logging
     common.configure_logging()