+++ gdm/test/test_plugin.py
@@ -1,4 +1,3 @@
-"""Unit tests for the 'plugin' module."""
 # pylint: disable=no-self-use
 
 from unittest.mock import patch, call
@@ -19,7 +18,7 @@ class TestMain:
 
         assert [
             call.install(root=None, depth=None,
-                         clean=False, force=False),
+                         clean=False, fetch=True, force=False),
             call.install().__bool__(),  # command status check
         ] == mock_commands.mock_calls