+++ gdm/test/test_shell.py
@@ -1,4 +1,3 @@
-"""Unit tests for the `shell` module."""
 # pylint: disable=no-self-use
 
 from unittest.mock import patch, Mock
@@ -6,7 +5,9 @@ from unittest.mock import patch, Mock
 import pytest
 
 from gdm.common import CallException
-from gdm.shell import call, ShellMixin, GitMixin
+from gdm import shell
+
+from . import assert_calls
 
 
 class TestCall:
@@ -16,190 +17,59 @@ class TestCall:
     @patch('os.chdir')
     def test_cd(self, mock_chdir):
         """Verify directories are changed correctly."""
-        call('cd', 'mock/dir')
+        shell.call('cd', 'mock/dir')
         mock_chdir.assert_called_once_with('mock/dir')
 
     @patch('gdm.shell.Command')
     def test_other(self, mock_command):
         """Verify directories are changed correctly."""
-        call('mock_program')
+        shell.call('mock_program')
         mock_command.assert_called_once_with('mock_program')
 
-    def test_other_error(self):
-        """Verify program errors are handled."""
-        with pytest.raises(SystemExit):
-            call('git', '--invalid-git-argument')
-
     def test_other_error_uncaught(self):
-        """Verify program errors can be left uncaught."""
+        """Verify program errors raise exceptions."""
         with pytest.raises(CallException):
-            call('git', '--invalid-git-argument', catch=False)
+            shell.call('git', '--invalid-git-argument')
 
     def test_other_error_ignored(self):
         """Verify program errors can be ignored."""
-        call('git', '--invalid-git-argument', ignore=True)
+        shell.call('git', '--invalid-git-argument', _ignore=True)
 
     def test_other_capture(self):
         """Verify a program's output can be captured."""
-        stdout = call('echo', 'Hello, world!\n', capture=True)
+        stdout = shell.call('echo', 'Hello, world!\n', _capture=True)
         assert "Hello, world!" == stdout
 
 
-def assert_calls(mock_call, expected):
-    """Confirm the expected list of calls matches the mock call."""
-    actual = [' '.join(args[0]) for args in mock_call.call_args_list]
-    assert expected == actual
-
-
 @patch('gdm.shell.call')
-class TestShell:
+class TestPrograms:
 
-    """Tests for calls to shell utilities."""
-
-    shell = ShellMixin()
+    """Tests for calls to shell programs."""
 
     def test_mkdir(self, mock_call):
         """Verify the commands to create directories."""
-        self.shell.mkdir('mock/dir/path')
+        shell.mkdir('mock/dir/path')
         assert_calls(mock_call, ["mkdir -p mock/dir/path"])
 
     def test_cd(self, mock_call):
         """Verify the commands to change directories."""
-        self.shell.cd('mock/dir/path')
+        shell.cd('mock/dir/path')
         assert_calls(mock_call, ["cd mock/dir/path"])
 
     @patch('os.path.isdir', Mock(return_value=True))
     def test_ln(self, mock_call):
         """Verify the commands to create symbolic links."""
-        self.shell.ln('mock/target', 'mock/source')
+        shell.ln('mock/target', 'mock/source')
         assert_calls(mock_call, ["ln -s mock/target mock/source"])
 
     @patch('os.path.isdir', Mock(return_value=False))
     def test_ln_missing_parent(self, mock_call):
         """Verify the commands to create symbolic links (missing parent)."""
-        self.shell.ln('mock/target', 'mock/source')
+        shell.ln('mock/target', 'mock/source')
         assert_calls(mock_call, ["mkdir -p mock",
                                  "ln -s mock/target mock/source"])
 
     def test_rm(self, mock_call):
         """Verify the commands to delete files/folders."""
-        self.shell.rm('mock/dir/path')
+        shell.rm('mock/dir/path')
         assert_calls(mock_call, ["rm -rf mock/dir/path"])
-
-
-@patch('gdm.shell.call')
-class TestGit:
-
-    """Tests for calls to Git."""
-
-    shell = GitMixin()
-
-    def test_clone(self, mock_call):
-        """Verify the commands to clone a new Git repository."""
-        self.shell.git_clone('mock.git', 'mock/path')
-        assert_calls(mock_call, ["git clone mock.git mock/path"])
-
-    def test_fetch(self, mock_call):
-        """Verify the commands to fetch from a Git repository."""
-        self.shell.git_fetch('mock.git')
-        assert_calls(mock_call, [
-            "git remote rm origin",
-            "git remote add origin mock.git",
-            "git fetch --tags --force --prune origin",
-        ])
-
-    def test_fetch_rev(self, mock_call):
-        """Verify the commands to fetch from a Git repository w/ rev."""
-        self.shell.git_fetch('mock.git', 'mock-rev')
-        assert_calls(mock_call, [
-            "git remote rm origin",
-            "git remote add origin mock.git",
-            "git fetch --tags --force --prune origin mock-rev",
-        ])
-
-    def test_fetch_rev_sha(self, mock_call):
-        """Verify the commands to fetch from a Git repository w/ SHA."""
-        self.shell.git_fetch('mock.git', 'abcdef1234' * 4)
-        assert_calls(mock_call, [
-            "git remote rm origin",
-            "git remote add origin mock.git",
-            "git fetch --tags --force --prune origin",
-        ])
-
-    def test_fetch_rev_revparse(self, mock_call):
-        """Verify the commands to fetch from a Git repository w/ rev-parse."""
-        self.shell.git_fetch('mock.git', 'master@{2015-02-12 18:30:00}')
-        assert_calls(mock_call, [
-            "git remote rm origin",
-            "git remote add origin mock.git",
-            "git fetch --tags --force --prune origin",
-        ])
-
-    def test_changes(self, mock_call):
-        """Verify the commands to check for uncommitted changes."""
-        self.shell.git_changes()
-        assert_calls(mock_call, [
-            # based on: http://stackoverflow.com/questions/3878624
-            "git update-index -q --refresh",
-            "git diff-index --quiet HEAD",
-            "git ls-files --others --exclude-standard",
-        ])
-
-    def test_changes_false(self, _):
-        """Verify the absence of changes can be detected."""
-        with patch('gdm.shell.call', Mock(return_value="")):
-            assert False is self.shell.git_changes()
-
-    def test_changes_true_untracked(self, _):
-        """Verify untracked files can be detected."""
-        with patch('gdm.shell.call', Mock(return_value="file_1")):
-            assert True is self.shell.git_changes()
-
-    def test_changes_true_uncommitted(self, _):
-        """Verify uncommitted changes can be detected."""
-        with patch('gdm.shell.call', Mock(side_effect=CallException)):
-            assert True is self.shell.git_changes()
-
-    def test_update(self, mock_call):
-        """Verify the commands to update a working tree to a revision."""
-        self.shell.git_update('mock_rev')
-        assert_calls(mock_call, [
-            "git stash",
-            "git clean --force -d -x",
-            "git checkout --force mock_rev",
-            "git branch --set-upstream-to origin/mock_rev",
-            "git pull --ff-only --no-rebase",
-        ])
-
-    def test_update_no_clean(self, mock_call):
-        self.shell.git_update('mock_rev', clean=False)
-        assert_calls(mock_call, [
-            "git stash",
-            "git checkout --force mock_rev",
-            "git branch --set-upstream-to origin/mock_rev",
-            "git pull --ff-only --no-rebase",
-        ])
-
-    def test_update_revparse(self, mock_call):
-        """Verify the commands to update a working tree to a rev-parse."""
-        mock_call.return_value = "abc123"
-        self.shell.git_update('mock_branch@{2015-02-12 18:30:00}')
-        assert_calls(mock_call, [
-            "git checkout --force mock_branch",
-            "git rev-list -n 1 --before='2015-02-12 18:30:00' mock_branch",
-            "git stash",
-            "git clean --force -d -x",
-            "git checkout --force abc123",
-            "git branch --set-upstream-to origin/abc123",
-            "git pull --ff-only --no-rebase",
-        ])
-
-    def test_get_url(self, mock_call):
-        """Verify the commands to get the current repository's URL."""
-        self.shell.git_get_url()
-        assert_calls(mock_call, ["git config --get remote.origin.url"])
-
-    def test_get_sha(self, mock_call):
-        """Verify the commands to get the working tree's SHA."""
-        self.shell.git_get_sha()
-        assert_calls(mock_call, ["git rev-parse HEAD"])